+++ source/manga-backend/common/src/main/java/com/raindrop/common/event/NewChapterEvent.java
@@ -13,6 +13,6 @@ public class NewChapterEvent {
     private String mangaId;
     private String mangaTitle;
     private String chapterId;
-    private int chapterNumber;
+    private double chapterNumber;
     private String chapterTitle;
 }
