+++ source/manga-backend/manga-service/src/main/java/com/raindrop/manga_service/dto/request/ChapterRequest.java
@@ -12,7 +12,7 @@ import java.util.List;
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class ChapterRequest {
-    int chapterNumber;
+    double chapterNumber;
     String title;
     List<MultipartFile> pages;
     String mangaId;
