+++ source/manga-backend/manga-service/src/main/java/com/raindrop/manga_service/entity/Chapter.java
@@ -21,7 +21,7 @@ public class Chapter {
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
     String id;
-    int chapterNumber;
+    double chapterNumber;
     String title;
     int views;
     int comments;
