+++ source/manga-backend/manga-service/src/main/java/com/raindrop/manga_service/repository/ChapterRepository.java
@@ -14,7 +14,7 @@ import java.util.Set;
 
 public interface ChapterRepository extends JpaRepository<Chapter, String> {
     Chapter findByTitle(String title);
-    Optional<Chapter> findByMangaAndChapterNumber(Manga manga, int chapterNumber);
+    Optional<Chapter> findByMangaAndChapterNumber(Manga manga, double chapterNumber);
     Set<Chapter> findByManga(Manga manga);
     List<Chapter> findByMangaId(String mangaId);
